// src/provider/client/models/user.ts
/* tslint:disable */
/* eslint-disable */
/**
 * KotiKota
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: latest
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import { BankInfo } from './bank-info';
import { CreateUser } from './create-user';
import { UserProfile } from './user-profile';
import { BankInfo, CreateUser, UserProfile } from '.';

/**
 *
 *
 * @export
 * @interface User
 */
export interface User extends CreateUser {
  /**
   * @type {UserProfile}
   * @memberof User
   */
  profile?: UserProfile;

  /**
   * @type {BankInfo}
   * @memberof User
   */
  bankInfo?: BankInfo;
}